Fix the ImplicitUnitMultipleOf canonicalizer rule
