+++ src/canonicalizer/CMakeLists.txt
@@ -17,6 +17,7 @@ add_library(sourcemeta_jsonbinpack_canonicalizer
   rules/implicit_object_properties.h
   rules/implicit_object_required.h
   rules/implicit_string_lower_bound.h
+  rules/implicit_unit_multiple_of.h
   rules/max_contains_without_contains.h
   rules/min_contains_without_contains.h
   rules/null_as_const.h
