+++ src/canonicalizer/canonicalizer.cc
@@ -24,6 +24,7 @@
 #include "rules/implicit_object_properties.h"
 #include "rules/implicit_object_required.h"
 #include "rules/implicit_string_lower_bound.h"
+#include "rules/implicit_unit_multiple_of.h"
 #include "rules/max_contains_without_contains.h"
 #include "rules/min_contains_without_contains.h"
 #include "rules/null_as_const.h"
@@ -52,6 +53,7 @@ sourcemeta::jsonbinpack::Canonicalizer::Canonicalizer(
   this->bundle.template add<ImplicitObjectProperties>();
   this->bundle.template add<ImplicitObjectRequired>();
   this->bundle.template add<ImplicitStringLowerBound>();
+  this->bundle.template add<ImplicitUnitMultipleOf>();
   this->bundle.template add<MaxContainsWithoutContains>();
   this->bundle.template add<MinContainsWithoutContains>();
   this->bundle.template add<NullAsConst>();
